server/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
groq
spacy
networkx
matplotlib
load-dotenv
easyocr
pillow 
pdf2image
pymupdf